Add Snyk
For the moment we use `pull_request` triggers, which only work with pull
requests from the main project (not forks).

We can later add `pull_request_target` to support PRs from forks, but
would want to use a pattern like
```
on:
  pull_request_target:
    types: [labeled]

jobs:
 snyk_vanagon:
   runs-on: ubuntu-latest
   if: contains(github.event.pull_request.labels.*.name, 'safe to test')
```
to ensure we don't run unauthorized code that could leak the Snyk token.